Upgrade scripts and new schemas for MySQL and PostgreSQL for 3.2.0

+++ genie-core/src/test/java/com/netflix/genie/core/jpa/entities/BaseEntityUnitTests.java
@@ -177,7 +177,7 @@ public class BaseEntityUnitTests {
     public void testEntityVersion() {
         final BaseEntity a = new BaseEntity();
         Assert.assertNull(a.getEntityVersion());
-        final Long version = 4L;
+        final Integer version = 4;
         a.setEntityVersion(version);
         Assert.assertEquals(version, a.getEntityVersion());
     }
